LPS-169905 Extract the logic of tracking class PersonalMenuEntry from class RolesAdminPortlet to class PersonalMenuEntryRegistryImpl
