+++ modules/apps/roles/roles-admin-web/src/main/java/com/liferay/roles/admin/web/internal/portlet/RolesAdminPortlet.java
@@ -22,9 +22,6 @@ import com.liferay.application.list.display.context.logic.PanelCategoryHelper;
 import com.liferay.application.list.display.context.logic.PersonalMenuEntryHelper;
 import com.liferay.depot.model.DepotEntry;
 import com.liferay.item.selector.ItemSelector;
-import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
-import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
-import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReferenceComparator;
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.portal.kernel.exception.DataLimitExceededException;
 import com.liferay.portal.kernel.exception.DuplicateRoleException;
@@ -68,6 +65,7 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.product.navigation.personal.menu.PersonalMenuEntry;
+import com.liferay.product.navigation.personal.menu.PersonalMenuEntryRegistry;
 import com.liferay.roles.admin.constants.RolesAdminPortletKeys;
 import com.liferay.roles.admin.constants.RolesAdminWebKeys;
 import com.liferay.roles.admin.panel.category.role.type.mapper.PanelCategoryRoleTypeMapper;
@@ -79,7 +77,6 @@ import com.liferay.segments.service.SegmentsEntryRoleLocalService;
 import java.io.IOException;
 
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -98,11 +95,7 @@ import javax.portlet.RenderResponse;
 import javax.portlet.ResourceRequest;
 import javax.portlet.ResourceResponse;
 
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceReference;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -347,7 +340,7 @@ public class RolesAdminPortlet extends MVCPortlet {
 	}
 
 	public List<PersonalMenuEntry> getPersonalMenuEntries() {
-		return _personalMenuEntryServiceTrackerList.toList();
+		return _personalMenuEntryRegistry.getPersonalMenuEntrys();
 	}
 
 	@Override
@@ -498,22 +491,6 @@ public class RolesAdminPortlet extends MVCPortlet {
 		}
 	}
 
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		Comparator<ServiceReference<PersonalMenuEntry>> groupComparator =
-			new PropertyServiceReferenceComparator<>(
-				"product.navigation.personal.menu.group");
-
-		Comparator<ServiceReference<PersonalMenuEntry>> entryOrderComparator =
-			new PropertyServiceReferenceComparator<>(
-				"product.navigation.personal.menu.entry.order");
-
-		_personalMenuEntryServiceTrackerList = ServiceTrackerListFactory.open(
-			bundleContext, PersonalMenuEntry.class,
-			Collections.reverseOrder(
-				groupComparator.thenComparing(entryOrderComparator)));
-	}
-
 	@Override
 	protected void checkPermissions(PortletRequest portletRequest)
 		throws Exception {
@@ -532,11 +509,6 @@ public class RolesAdminPortlet extends MVCPortlet {
 		}
 
 		super.checkPermissions(portletRequest);
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_personalMenuEntryServiceTrackerList.close();
 	}
 
 	@Override
@@ -855,8 +827,8 @@ public class RolesAdminPortlet extends MVCPortlet {
 	private PanelCategoryRoleTypeMapperRegistry
 		_panelCategoryRoleTypeMapperRegistry;
 
-	private ServiceTrackerList<PersonalMenuEntry>
-		_personalMenuEntryServiceTrackerList;
+	@Reference
+	private PersonalMenuEntryRegistry _personalMenuEntryRegistry;
 
 	@Reference
 	private Portal _portal;
